make clickable SelectionListItem s focusable
I need this so the SearchAutocompleteDropdown results can be clickable,
while still hiding the autocomplete results if you focus something
outside of SearchBar and its descendants.

This is also an accessibility improvement, compared to having a list
item (<li>) with an onClick. However, it actually makes an accessibility
lint warning complain, because technically <button> is more semantic
than <a>. While this is true, resetting browser default styles for
<button> is a pain in the butt, and I don't have the energy to take it
on right now.