fix some stuff but still not working in iOS safari
this corrects an error where the autocomplete results were not a child
of the scrollable element.

this makes it work right in the simulator when I manually set body
height to 280px. I do not know why it's still broken on mobile safari.

brokenness on mobile safari:

 - cannot scroll the bottom pane up or down
 - I guess that's it actually. Manifests as both not being able to
scroll up and see the padding on start screen/routes overview, and also
not being able to scroll the autocomplete list; the results scroll
instead.