Fix autocomplete results flickering
A while back I wrote code to try to provide more useful autocomplete
results by looking back: If we have autocomplete results for "Glen Ca"
but no results for the current contents of the input box which are
"Glen Cany", remove letters until we do find results so that Glen
Canyon Park can be displayed as a suggestion.

This was never quite working though, because it only tried removing
letters if there was no cache record at all, but gave up immediately
if there was a cache record in the "fetching" state. In other words,
this set of events could happen:

1. Type "Glen Ca", pause for autocomplete results for "Glen Ca" to
come back
2. Type "ny". Since there is no cache record for "Glen Cany", we try
removing letters, and you still get the same results as in step 1
3. After the 700ms debounce delay, we kick off a fetch for "Glen Cany"
and create a cache record for this key with status "fetching". Your
search results now DISAPPEAR
4. Finally, autocomplete results for "Glen Cany" come back, and the
entry for Glen Canyon Park, having disappeared, reappears