iPhone home screen icon fix + theme color
If you add BikeHopper to your home screen on an iPhone, the icon is
just black because it's a black line drawing of a bike on a
transparent background and iOS puts that on a black background. To
work around this, I've made the large sizes of the icon explicitly
have a white background, but have kept favicon.ico with a transparent
background so browser tabs on desktop will look good.

Also, updated the theme_color parameter in the manifest to match that
in the index.html