Add padding to ensure bottom-of-screen content is clickable
Where possible, we use env(safe-area-inset-bottom) to add only the
padding necessary.

On mobile Safari, if env(safe-area-inset-bottom) is 0, we assume it's
lying and add 74px of padding. I (Scott) have determined this is
necessary on my iPhone SE 1st gen running iOS 15. This workaround is
only given to Safari, and not when running as a PWA (ie saved to home
screen).

To make it look less janky with the padding, the last route in the
routes overview now has a bottom border.