Fix scroll to bottom on Android first load
Existing code, when the Visual Viewport height changes, updates the body
height to match the Visual Viewport height to prevent vertical
scrolling.

This commit also sets the body height to the Visual Viewport height,
if supported, on first load. This fixes a bug observed on Android
Chrome where, on first load (but not after going into the location
inputs and back), it was impossible to scroll the bottom drawer's
content all the way to the bottom.