Reduce jank when entering location inputs on iOS
On mobile Safari, when you tap into the decoy destination input at the
bottom of the screen, Safari tries to scroll down to that input.
Existing code hacks around this by shrinking the body height to
reflect the virtual keyboard being displayed, but this causes the
browser to animate scrolling down and then scrolling back up.

This commit makes that glitch less noticeable by *also* scrolling up
immediately in response to the decoy input's focus event.