Let bottom drawer be resized smaller
Existing code makes the bottom drawer's minimum and default height both
180px. On a phone with limited real estate, such as my 1st gen iPhone
SE, this can be bigger than you might want. It doesn't give you a lot of
room to see the map.

(Recall that the bottom drawer is implemented by having a combination of
a transparent div (through which the map can be seen and interacted
with) and a div for the bottom content, which together scroll
vertically. Thus, making the bottom drawer's minimum height 180px means
making the transparent div's minimum height "calc(100% - 180px)".)

This commit makes the minimum height 140px and the default height 200px.
It does this by making the transparent div in front of the map be
"calc(100% - 140px)" tall -- allowing the minimum height of the bottom
drawer to be 140px, as desired -- and then setting the initial scrollTop
of the scrolling container to 60px, the difference between the desired
default and minimum bottom drawer height.

Known bug: On Android, for some reason, the call to center the map
around a displayed route does not appear to be taking the 60px
difference into account, even though the BikeHopper component explicitly
does add that to the bottom padding. I will look into this later. It's
easily worked around (just resize the bottom drawer smaller) and this
change is a significant win for iOS.