Make transitions use default bike color (asphalt gray)
We never explicitly decided on a color for transition curves between
legs, and were falling back to the green BikeHopper theme color until
the pull request #109 removed that default, leading to a console warning
from MapLibre and defaulting to black.

For lack of a better idea, I'm specifying the transition curves will be
asphalt gray, same as used for bike routing on streets without bike
infrastructure.