Limit map to bounds of data cutout
Our data cutout is actually non-rectangular:

https://download.geofabrik.de/north-america/us/california/norcal.html

For now I'm erring on the permissive side and using a bounding box
around this cutout, so it will be possible to request routes to/within
Nevada that our GraphHopper server won't actually route to. (As well
as large areas of NorCal that are not the Bay Area where we can do
street routing but have no public transit data.)

Close enough. This avoids giving the false impression that you could
use our BikeHopper instance to route you around Portland or
Washington, DC or London.