Warn when local transit data is missing (#209)
This warns you when your route is likely missing transit options,
because it's within our road service area but outside our transit
service area.

Fixes #164.

The approach is that we manually run a script that generates a bounding
polygon around all local transit stops. (More below on bounds.) We then
have to paste the output of that script into the Bay Area regional
definitions file and commit the result.

It would obviously be cleaner to automatically run this on the server
side each time new GTFS data is fetched. But that's architecturally more
complicated. That can be future work if we want.

The script takes about 45 seconds on my laptop. It's written naively and
could probably be optimized to take 1/10th that time or less, but also,
who cares.

When porting BikeHopper to a new region, implementing this is optional.
If the TRANSIT_SERVICE_AREA constant is null, then BikeHopper will
simply never display this warning. The algorithm should be reasonable to
apply to any region, although a few transit lines/agencies are excluded
and that's Bay Area specific: The bounds exclude Amtrak, Capital
Corridor, and the SolTrans B line, an outlier route which goes to Davis
and Sacramento.