Show correct transit type & icon (bus, tram, train, etc) (#215)
Up until now we have always shown a bus icon for every transit leg,
even if it is BART, Muni light rail, a ferry, etc., and we've used the
generic word "line" ("Ride the N line").

The route type is actually provided for us in GTFS though. See the
spec: https://gtfs.org/schedule/reference/#routestxt

But GraphHopper hasn't been patching route_type through in its
responses. That will be fixed by
https://github.com/bikehopper/graphhopper/pull/111

This commit uses the route_type, if present, to display the correct
icon and a more specific noun ("Ride the N train"). When this and the
above router change are both merged and deployed, this will fix #206.