Fix routes sometimes not loading from URL
I noticed sometimes the route params in the URL were ignored

This seems to be happening because the URL middleware, where this
initialization happens, doesn't get called on the Redux @@init event,
and then there isn't any other event.

This works around the issue by firing an otherwise ignored map_loaded
event when the map loads.