Restore last viewed routes when installed as a PWA
When running as a Progressive Web App (i.e. installed to Home Screen),
we haven't been restoring the last viewed routes after the app pages
out of memory, as we do when running within a Safari tab.

This, I believe, is because the PWA manifest tells the OS to always
start the app at the same URL, in our case the root URL, so it doesn't
have the path with the route params in it.

This commit works around that by copying the path into localStorage.
Tested on my iPhone X. Fixes #226