Use custom place icons in search bar
Status quo: If we have a relevant custom icon for a feature (e.g.
steaming coffee mug for a cafe, tree icon for a park), we display that
icon in the autocomplete drop down instead of a pin. But we still always
use the generic pin icon in the search bar itself.

This commit: Uses the custom icon, if any, in the search bar itself for
geocoded features. It will be replaced with a generic pin icon if the
location text is subsequently edited.

Limitation of this commit: If the location is hydrated from the URL
(reload page or follow link), the custom icon will be lost because in
that case we only have the place name and coordinates, not further
details.