Make fitting map to route bounds more reliable
Fixes #139

Existing code: Measures the distance between the bottom of the map
overlay's transparent window (.MapPlusOverlay_window) and the bottom
of the browser window (i.e. window.innerHeight) to decide how much
padding to add when zooming the map.

The default scroll (i.e. the difference between the default height of
the bottom drawer and the minimum height of the bottom drawer) is
added to that value. This essentially assumes that the map zooming
code is running after the bottom drawer is rendered but before the
default scroll is applied (in `<MapPlusOverlay>`).

But there is no guarantee the code runs in that order. The map zooming
code can run either synchronously or asynchronously, after waiting for
evidence that the mobile device's virtual keyboard has gone down.

Existing code also calls .resize() on the map before fitting bounds,
but this has not been necessary for a long time, since we now have the
map object fill the browser window, and layer other UI elements on top
of it so it doesn't have to resize.

This commit: Instead of measuring the bottom drawer's actual
dimensions, uses the hardcoded constants for the bottom drawer's
default height to calculate padding. This means it doesn't matter
whether it runs before or after the bottom drawer is rendered or the
default scroll is applied. It still uses measurement to get the height
of the top bar, but that's OK because the top bar height isn't
changing while routes are being fetched and displayed.

This also changes the `overlayRef` prop to `<BikehopperMap>` to be the
same as the `mapOverlayRef` variable in `<MapPlusOverlay>`. It kept
confusing me that those were different.

Lastly, this removes the unnecessary call to .resize().