Simplify SelectionListItem + fix rounded borders (#232)
This is a refactor, small a11y improvement, and small visual
enhancement, motivated by work I'm doing on a different branch but
reviewable on its own.

Motivation: I'm adding storage of recently used locations so you can
easily find directions again to a place you just looked up. Thus, I
want to add an X-out so you can delete a place you don't want showing
up in that list.

This commit doesn't do that however. It:

1. Simplifies SelectionListItem, which was using two different
versions of markup depending on whether the item was clickable or not.
But it's a selection list, so the items are always supposed to be
clickable.

2. Makes SelectionListItem use a <button> instead of a link. This is
better for a11y.

3. Removes the one non-clickable usage of SelectionListItem: the
warning for missing transit info (Warn when local transit data is
missing #209).

4. The only real user-visible change: Puts rounded borders at the top
of the routes overview, which matches the rounded borders on the
initial screen and on the itinerary view. I just realized that the
rounded borders were supposed to be there and were being inadvertently
hidden.