Change how transit modes are handled; partly internationalize
Internationalizing requires that we stop passing around transit modes
as strings such as "bus" and "train" that we directly plug into
sentences. Instead, we need codes representing the mode, and we should
wait to convert that to end-user text until we render the mode in a
specific piece of UI.

There is no reason not to just use the GTFS enum as that code. So now,
instead of having a lookup table from GTFS enum to string and icon, we
have a constant table that makes the GTFS enum easier to use
(lib/TransitModes, similar to the existing lib/InstructionSigns), and
a <ModeIcon> component that renders an icon based on a GTFS enum
value (sort of like <PlaceIcon>, but slightly different as noted in
<ModeIcon> comments).

The alt text label for the mode description in the routes overview is
internationalized, but the itinerary header text is not yet, and is
temporarily broken (It currently shows the number, i.e. "Ride the
Red-S 1 (BART)"). Will fix in followup commit.