Switch from create-react-app to Vite
We've talked about using a component library like Radix, designed to be
used with TailwindCSS. In turn, TailwindCSS [advises against using
create-react-app](https://tailwindcss.com/docs/guides/create-react-app)
due to lack of support for tools in the PostCSS ecosystem and recommends
[Vite](https://vitejs.dev/), Parcel, Next.js, or Remix.

I took a look at Vite and it seems good and was a relatively easy
replacement. Just had to find the right plugins, rename .js -> .jsx for
files with JSX markup, and rename env vars. As a bonus, all the crap we
ended up with in the repo when we ejected from create-react-app can be
deleted.

After this change, to keep your dev server working you'll have to:

1. Edit .env.local.development, rename everything prefixed with
   REACT_APP_ to be prefixed with VITE_
2. Re-accept the self-signed SSL cert in your browser, as it'll be a
   different one