Fix serial selection of different autocompletes
Fixes #243. Restating the repro steps described there:

1. Type "reveille," all the different Cafe Reveille locations come up
2. Select the top one [say Polk St]. Then tap into destination box
   again. You get the same list of Reveilles again.
3. Select the second from top [say Long Bridge St]. Tap into
   destination box again.

You should get the same list a third time and be able to select a
third Reveille location and so on, but there were no results. This
commit fixes this issue.

The key is that the "fromInputText" attribute in the "select geocoded
location" action should not necessarily be the *current* input text,
but rather the original input text that produced the list of geocoded
locations being selected from.