Store OSM IDs with type attached
Geocoding results can be a node, way, or relation. Just using the OSM ID
alone is ambiguous, as the node with ID 100 (say) has nothing to do with
the way with ID 100, etc.

This commit solves the issue by using strings that begin with "N", "W",
or "R", followed by a stringified ID, everywhere OSM IDs are used as a
key. The format used to save recently-used locations in localStorage
does not change. Fixes #302