After permission denied, don't default to geolocation
Usually, BikeHopper defaults to geolocating your start location when you
enter an end location (or select an end location on the map).

This commit makes it so that after geolocation fails due to a permission
denied error, we will no longer attempt to default to it. We will just
leave the start location field blank in those cases and let you manually
select.

We'll restore the behavior of defaulting to geolocating your start
location after the page is reloaded or after geolocation is manually
selected and succeeds.

Fixes #342