Fix linting of TypeScript files
`npm run lint` was not scanning ts and tsx files.

This also turns off TS's unused locals/args warning since it is
duplicative of this more configurable lint rule:

https://typescript-eslint.io/rules/no-unused-vars

I have not turned on the Typescript-ESLint recommended rules yet because
it produces lots of errors that will need to be dealt with in time.