Code split gpx2fit
gpx2fit is a relatively large dependency (~300kB uncompressed) that many
BH users will not use. This uses a dynamic import to defer loading it
until a FIT file is downloaded.

To show it's doing something, the FIT button grays while the code is
being downloaded, and if it fails to load, an alert is displayed. I
tested this by adding a delay and by adding a throw, respectively.