Fix scroll to step on return from single-step view
(and expanding transit legs on return from single-stop view)

The leg that contains the step to be scrolled to 1) must be expanded,
and 2) this must happen on the first render, so it can't be delayed by a
useEffect. We also don't need the useEffect because this currently never
needs to be recomputed without the component unmounting and being
recreated.