tabindex hack to attempt to fix dropdown accessibility
It sort of works except that after the end of the dropdown, tab order
goes, not to the next element in the SearchBar (that would be the
TimeBar's button), but to the map, which is invisible, and then to each
of its controls in turn. Adding aria-hidden to the map did not affect
this. I'd have to also set tabIndex={-1} on the map and each of its controls
and it looks like ReactMapGL doesn't even have a mechanism for this.