Better dropdown options after picking recently used
There was a bug where if you select a location from the recently-used
list, then go back into the relevant input box, the dropdown is empty,
and spins endlessly. Now it will instead provide other recently-used
options and/or "Current Location", if those are not used in the other
input box.