Improve start/end editing with UrlWithString source
A location with a UrlWithString source means that the location came from
geocoding, but the page was subsequently reloaded or passed around as a
link.

Before this commit, when focusing a search box with a UrlWithString
location, no dropdown options are provided. This provides "Current
Location" and recently-used options, for parity with locations with
Geocoded source.