fix scrolling
I added position: relative to the main div of ItinerarySingleTransitStop
to position the X in the top right corner. But this made that div the
offsetParent, so useScrollToRef tried to scroll within it, instead of
the actual scroll container (div.MapPlusOverlay_overlay). So I'm
modifying useScrollToRef to iterate till it finds an actually scrollable
container.